Make calls to get never delete items even if they are expired
An alternative to 495e2b7481d4a3255f92931484318656dc157a9d

This fixes an issue where you could

if (cache.has(key)) {
	const value = cache.get(key)
	// assume that value is a legit value
}

and have the call to cache.get() return undefined even though you just checked that it was there by calling cache.has().

It seems counterintuitive (and counter to the docs) that a call to get() would ever remove an item from the cache.